backend/requirements.txt
Flask==2.3.2
flask-cors==6.0.0
flask_caching==2.3.1
pandas==2.0.3
openpyxl==3.1.2
pymysql==1.0.3
pymysqlpool==1.0.9
DBUtils==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.3
pyjwt==2.8.0
qrcode==7.4.2
openai==1.106.1
psycopg2-binary==2.9.6
gunicorn==21.2.0
numpy==1.26.4
matplotlib==3.7.2
docx==0.2.4
python-docx==1.2.0
pymupdf==1.26.0
redis==7.0.1